refactor(configuration): вынесены утилиты для работы с файлами версий конфигурации
- Добавлен новый файл `configVersionUtils.ts` с функциями `checkVersionFileExists`, `handleMissingVersionFile` и `clearDirectory` для проверки и обработки файлов версий конфигурации.
- Обновлен класс `ConfigurationCommands` для использования новых утилит вместо встроенных методов.
- Упрощен код и улучшена читаемость, добавлены соответствующие JSDoc комментарии для новых функций.
